PROTOCOL-117 swap SearchResponseEncoderTest and ListResponseEncoderTest (no code change)
